// src/pages/Settings/Companies/types.ts
export type Company = {
    id: number;
    name: string;
    director_name: string;
    tin: string;
    mfo: string;
    account_number: string;
    phone_number: string;
    address: string;
    bank_name: string;
    region_id: number;
    district_id: number;
    oked: string;
    status: number;
    is_active: number;
    created_at: string;
};